+++ Expanded Cheat Menu/Changelog.md
@@ -32,6 +32,9 @@
 - > It'll try to list files in mount points (ie: ~"Textures" or ~"Prefabs"), needs the HelperMod installed.
 - > The X close button now checks if Ctrl or Shift is being held down (see tooltip for more info).
 - > Merged View/Dump menu items in Tools.
+- Console Window:
+- > Renamed Copy Log Text to View Log Text.
+- > Added a Clipboard button (copies log to clipboard).
 
 ### Fixed:
 - Game would freeze if ConsoleErrors was false (thanks Gaspurr).